test: 입력 검증 테스트 코드 및 파일명 변경
